requirements.txt
asgiref==3.2.7
Django==3.0.7
django-cors-headers==3.3.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
